Split QueryStore History + Overview controls into partials
Both controls were unsplit god files — QueryStoreHistoryControl in
particular was the team's outlier, while every comparable sibling control
already uses feature partials. Bring them in line (pure mechanical moves,
no behavior change):

QueryStoreHistoryControl.axaml.cs (1104 -> 321):
  .Fetch.cs       — LoadHistoryAsync
  .Grid.cs        — color-indicator plumbing + grid<->chart selection sync
  .Legend.cs      — legend panel + plan highlighting
  .Chart.cs       — chart drawing, smart X axis, dot highlighting
  .Selection.cs   — pointer/box selection + hover tooltip
  .PlanLoad.cs    — context menu + load-plan-from-selection

QueryStoreOverviewControl.axaml.cs (823 -> 250):
  .Donut.cs          — donut/arc geometry + popup + legend
  .WaitStatsChart.cs — stacked wait-stats chart
  .BarCards.cs       — metric bar cards (+ shared _dbColorMap)

Section divider comments travel with their sections.
